+++ loki-support/loki-support-redis/src/main/java/io/github/guoshiqiufeng/loki/support/redis/config/RedisAutoConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2023, fubluesky (fubluesky@foxmail.com)
+ * Copyright (c) 2023-2024, fubluesky (fubluesky@foxmail.com)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,8 +15,6 @@
  */
 package io.github.guoshiqiufeng.loki.support.redis.config;
 
-import io.github.guoshiqiufeng.loki.enums.MqType;
-import io.github.guoshiqiufeng.loki.support.core.config.GlobalConfig;
 import io.github.guoshiqiufeng.loki.support.core.config.LokiProperties;
 import io.github.guoshiqiufeng.loki.support.redis.RedisClient;
 import io.github.guoshiqiufeng.loki.support.redis.impl.RedisClusterImpl;
@@ -30,10 +28,7 @@ import org.springframework.context.annotation.Configuration;
 import redis.clients.jedis.*;
 
 import java.util.Arrays;
-import java.util.List;
 import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 /**
@@ -132,7 +127,7 @@ public class RedisAutoConfiguration {
     @Bean
     @ConditionalOnMissingBean(RedisClient.class)
     public RedisClient redisClient(LokiProperties lokiProperties, RedisProperties redisProperties) {
-        convert(lokiProperties, redisProperties);
+        RedisConfigUtils.convert(lokiProperties, redisProperties);
         if (redisProperties.getSentinel() != null && !redisProperties.getSentinel().getMaster().isEmpty()) {
             // 如果配置了 Sentinel，返回 RedisSentinelImpl 对象
             return new RedisSentinelImpl(jedisSentinelPool(redisProperties));
@@ -142,133 +137,6 @@ public class RedisAutoConfiguration {
         } else {
             // 默认情况下返回 RedisDefaultImpl 对象
             return new RedisDefaultImpl(redisPoolFactory(redisProperties));
-        }
-    }
-
-    /**
-     * 转换 Loki 配置和 Redis 配置，更新 RedisProperties 对象。
-     *
-     * @param lokiProperties  Loki 配置属性
-     * @param redisProperties Redis 相关配置属性
-     */
-    private void convert(LokiProperties lokiProperties, RedisProperties redisProperties) {
-        // 获取 Loki 全局配置
-        GlobalConfig globalConfig = lokiProperties.getGlobalConfig();
-        GlobalConfig.MqConfig mqConfig = globalConfig.getMqConfig();
-
-        if (isRedisMqConfigValid(mqConfig)) {
-            String address = mqConfig.getAddress();
-            if (address != null && !address.isEmpty()) {
-                // 将地址拆分为集合
-                List<String> addressList = Arrays.asList(address.split(","));
-                String sentinelMaster = getSentinelMaster(address);
-                boolean isSentinel = !sentinelMaster.isEmpty();
-                if (addressList.size() == 1) {
-                    // 单节点配置
-                    configureSingleNode(addressList.get(0), redisProperties);
-                } else if (addressList.size() > 1 && !isSentinel) {
-                    // 集群配置
-                    configureCluster(addressList, redisProperties);
-                } else if (addressList.size() > 1) {
-                    // Sentinel 配置
-                    addressList = addressList.stream().filter(tmp -> !tmp.equals(sentinelMaster))
-                            .collect(Collectors.toList());
-                    configureSentinel(addressList, redisProperties, mqConfig, sentinelMaster);
-                }
-            }
-
-            // 设置其他相关配置，如 auth、username、password 等。
-            if (mqConfig.getAuth() != null) {
-                configureAuth(mqConfig.getUsername(), mqConfig.getPassword(), redisProperties);
-            }
-        }
-    }
-
-    /**
-     * 检查 Redis Mq 配置是否有效。
-     *
-     * @param mqConfig Redis Mq 配置
-     * @return 配置是否有效
-     */
-    private boolean isRedisMqConfigValid(GlobalConfig.MqConfig mqConfig) {
-        return mqConfig != null && MqType.REDIS.equals(mqConfig.getMqType());
-    }
-
-    /**
-     * 根据单节点配置更新 RedisProperties 对象。
-     *
-     * @param node            单节点地址
-     * @param redisProperties Redis 相关配置属性
-     */
-    private void configureSingleNode(String node, RedisProperties redisProperties) {
-        String[] nodeArr = node.split(":");
-        if (nodeArr.length == 1) {
-            redisProperties.setHost(nodeArr[0]);
-        } else if (nodeArr.length == 2) {
-            redisProperties.setHost(nodeArr[0]);
-            redisProperties.setPort(Integer.parseInt(nodeArr[1]));
-        }
-    }
-
-    /**
-     * 根据集群配置更新 RedisProperties 对象。
-     *
-     * @param nodes           集群节点列表
-     * @param redisProperties Redis 相关配置属性
-     */
-    private void configureCluster(List<String> nodes, RedisProperties redisProperties) {
-        redisProperties.setCluster(new RedisProperties.Cluster().setNodes(nodes));
-    }
-
-    /**
-     * 根据 Sentinel 配置更新 RedisProperties 对象。
-     *
-     * @param nodes           Sentinel 节点列表
-     * @param redisProperties Redis 相关配置属性
-     * @param mqConfig        Redis Mq 配置
-     * @param master          Sentinel 主节点名称
-     */
-    private void configureSentinel(List<String> nodes, RedisProperties redisProperties,
-                                   GlobalConfig.MqConfig mqConfig, String master) {
-        // 设置 Sentinel 配置
-        RedisProperties.Sentinel sentinel = new RedisProperties.Sentinel()
-                .setNodes(nodes)
-                .setMaster(master);
-        if (mqConfig.getUsername() != null) {
-            sentinel.setUsername(mqConfig.getUsername());
-        }
-        if (mqConfig.getPassword() != null) {
-            sentinel.setPassword(mqConfig.getPassword());
-        }
-        redisProperties.setSentinel(sentinel);
-    }
-
-    /**
-     * 从 Sentinel 地址中获取主节点名称。
-     *
-     * @param address Sentinel 地址
-     * @return Sentinel 主节点名称
-     */
-    private String getSentinelMaster(String address) {
-        String sentinelPattern = "(\\w+)";
-        Pattern pattern = Pattern.compile(sentinelPattern);
-        Matcher matcher = pattern.matcher(address);
-        return matcher.find() ? matcher.group(1) : "";
-    }
-
-    /**
-     * 根据认证信息更新 RedisProperties 对象。
-     *
-     * @param username        认证用户名
-     * @param password        认证密码
-     * @param redisProperties Redis 相关配置属性
-     */
-    private void configureAuth(String username, String password, RedisProperties redisProperties) {
-        if (username != null) {
-            redisProperties.setUsername(username);
-        }
-        if (password != null) {
-            redisProperties.setPassword(password);
         }
     }
