+++ spring-boot-starter/loki-spring-boot-autoconfigure/src/main/java/io/github/guoshiqiufeng/loki/autoconfigure/config/KafkaAutoConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2023, fubluesky (fubluesky@foxmail.com)
+ * Copyright (c) 2023-2024, fubluesky (fubluesky@foxmail.com)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
